requirements.txt
asgiref==3.9.1
attrs==26.1.0
autobahn==25.12.2
automat==25.4.16
cbor2==5.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.2.1
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.5
cssbeautifier==1.15.4
daphne==4.2.1
django==5.2.5
django-eventstream==5.3.3
django-grip==3.5.2
djlint==1.36.4
editorconfig==0.17.1
gripcontrol==4.2.1
hyperlink==21.0.0
idna==3.11
incremental==24.11.0
jsbeautifier==1.15.4
json5==0.12.1
markupsafe==3.0.3
msgpack==1.1.2
packaging==26.0
pathspec==0.12.1
pillow==12.1.1
pip==25.1.1
pubcontrol==3.5.0
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pyjwt==2.12.1
pyopenssl==26.0.0
python-decouple==3.8
pyyaml==6.0.2
redis==7.3.0
regex==2025.7.34
requests==2.32.5
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
twisted==25.5.0
txaio==25.12.2
typing-extensions==4.15.0
ujson==5.12.0
urllib3==2.6.3
werkzeug==3.1.7
zope-interface==8.2
psycopg2-binary==2.9.10
